created can.d.ts to start typing out can
